requirements.txt
Flask==2.3.2
feedparser==6.0.8
python-telegram-bot==13.7
discord.py==1.7.3
requests